fetcher_service/requirements.txt
requests==2.25.1
kafka-python==2.0.2
